[TASK] moved tune to appropriate space
